requirements.txt
accelerate==0.24.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
bentoml==1.1.8
bitsandbytes==0.41.1
build==1.0.3
cattrs==23.1.2
certifi==2023.7.22
charset-normalizer==3.3.2
circus==0.18.0
click==8.1.7
click-option-group==0.5.6
cloudpickle==3.0.0
coloredlogs==15.0.1
contextlib2==21.6.0
dataclasses-json==0.6.1
datasets==2.14.6
deepmerge==1.1.0
Deprecated==1.2.14
dill==0.3.7
distro==1.8.0
dnspython==2.4.2
fastcore==1.5.29
filelock==3.13.1
filetype==1.2.0
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.10.0
ghapi==1.0.4
h11==0.14.0
httpcore==1.0.1
httpx==0.25.1
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
inflection==0.5.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.331
langsmith==0.0.59
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.1
openai==0.28.1
openllm==0.3.14
openllm-client==0.3.14
openllm-core==0.3.14
opentelemetry-api==1.20.0
opentelemetry-instrumentation==0.41b0
opentelemetry-instrumentation-aiohttp-client==0.41b0
opentelemetry-instrumentation-asgi==0.41b0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
opentelemetry-util-http==0.41b0
optimum==1.14.0
orjson==3.9.10
packaging==23.2
pandas==2.1.2
pathspec==0.11.2
Pillow==10.1.0
pinecone-client==2.2.4
pip-requirements-parser==32.0.1
pip-tools==7.3.0
prometheus-client==0.18.0
protobuf==4.25.0
psutil==5.9.6
pyarrow==14.0.0
pydantic==1.10.13
pydantic_core==2.10.1
Pygments==2.16.1
pynvml==11.5.0
pyparsing==3.1.1
pyproject_hooks==1.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
regex==2023.10.3
requests==2.31.0
rich==13.6.0
safetensors==0.4.0
schema==0.7.5
sentencepiece==0.1.99
simple-di==0.1.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.23
starlette==0.32.0.post1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tiktoken==0.5.1
tokenizers==0.14.1
torch==2.1.0
tornado==6.3.3
tqdm==4.66.1
transformers==4.35.0
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
uvicorn==0.24.0.post1
watchfiles==0.21.0
wcwidth==0.2.9
wrapt==1.15.0
xxhash==3.4.1
yarl==1.9.2
zipp==3.17.0
